WEBOPS-791: exhibitor watcher
Query an exhibitor and configure Kafka to use the returned ensemble.
Continue polling exhibitor, and if the ensemble changes:
- Acquire a lock in ZK (to prevent every Kafka instance restarting at
  once).
- Restart Kafka with the new ensemble.
- Wait around after the restart for sync/just in case.
- Release lock.

Ignore when the ZK ensemble shrinks. That means in a typical ASG health
check replacement, Kafka is only restarted _after_ the replacement is
online (there's no sense going from 2/3 to 2/2 healthy ZKs)